requirements.txt
Flask
gunicorn
PyPDF2
pytesseract
pdf2image
Pillow
python-dotenv
gevent
psutil

